skill_identifier/requirements.txt
wheel
ninja
packaging
torch==2.3.1
datasets==3.0.2
transformers==4.46.0
accelerate
sentencepiece
flash-attn==2.6.3
pytrec_eval
rouge_score
pyyaml
google==3.0.0
protobuf==4.25.5
pytest
vllm
mamba-ssm==2.2.2
causal-conv1d>=1.2.0
openai==1.52.2
anthropic==0.37.1
google-generativeai==0.7.2
vertexai==1.64.0